the max in min value for the Joystick now come from the function setJoystcik
Add two more arguments for the fucktion setJoystick to make it work with esp32 and the arduino.
the analoge read for the esp32 is between 0 an 4095.
the analoge read for the arduino is bewenn 0 an 1024.
if now values are past to the function it will use the default values.
the default values of this function are set to be compatible with the arduino.